project structure set up the way I like it
