+++ src/app/App.tsx
@@ -1,0 +1,13 @@
+import "../styles/App.css";
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <p>Home page</p>
+      </header>
+    </div>
+  );
+}
+
+export default App;
